app/requirements.txt
Django==3.0.7
gunicorn==20.0.4
psycopg2-binary==2.8.5
django-json-widget==1.0.1
djangorestframework==3.12.1
